+++ packages/coins-sdk/src/index.ts
@@ -23,6 +23,9 @@ export {
 } from "./actions/updatePayoutRecipient";
 export type { UpdatePayoutRecipientArgs } from "./actions/updatePayoutRecipient";
 
+export { tradeCoin, createTradeCall } from "./actions/tradeCoin";
+export type { TradeParameters } from "./actions/tradeCoin";
+
 // API Read Actions
 export * from "./api/queries";
 export type * from "./api/queries";
@@ -36,3 +39,6 @@ export { setApiKey } from "./api/api-key";
 
 // Metadata Validation Utils
 export * from "./metadata";
+
+// Uploader
+export * from "./uploader";